Correcao de Quadro de Superavit

+++ services/orcamento/ValidaSuplementacaoService.php
@@ -132,7 +132,7 @@ class ValidaSuplementacaoService
      */
     public function pegarValorQuadroSuperavitDeficit($sFonte)
     {
-        return (float) number_format($this->oQuadroSuperavitDeficit->pegarValorPorFonte($sFonte), 2, ".", "");
+        return number_format($this->oQuadroSuperavitDeficit->pegarValorPorFonte($sFonte), 2, ".", "");
     }
 
     /**
